chore: rename years.ts to dates.ts
The file is renamed to better reflect its expanded scope,
as it is going to support comprehensive date conversions,
not just year-specific functionality.
